Cambio noticia intento integracion, fallido por ahora

+++ Backend/eTours.logic/src/main/java/co/edu/uniandes/csw/eTours/Paquete/entity/PaqueteEntity.java
@@ -13,7 +13,7 @@ public class PaqueteEntity {
     @Id
     @GeneratedValue(generator = "Paquete")
     private Long id;
-    private Double descount;
+    private int descount;
     private String servicios;
 
      
@@ -27,11 +27,11 @@ public class PaqueteEntity {
         this.id = id;
     }
 
-    public Double getDescuento() {
+    public int getDescuento() {
         return descount;
     }
 
-    public void setDescuento(Double descuento) {
+    public void setDescuento(int descuento) {
         this.descount = descuento;
     }
